requirements.txt
aiogram>=3.0.0
python-dotenv
requests
fpdf2
tabulate
currency_converter_free
pyyaml
pydantic
aiohttp
